by 1113

+++ ft_printf/libftprintf/ft_putnbr.c
@@ -6,13 +6,13 @@
 /*   By: jyamada <jyamada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/21 21:19:41 by jyamada           #+#    #+#             */
-/*   Updated: 2025/11/13 19:26:10 by jyamada          ###   ########.fr       */
+/*   Updated: 2025/11/13 21:06:28 by jyamada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(size_t n)
 {
 	long	longn;
